requirements.txt
Django==5.2
djangorestframework==3.16.0

firebase-admin==6.8.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-auth==2.40.3
google-api-core==2.25.0
google-api-python-client==2.171.0

psycopg2==2.9.10
pillow==11.2.1
requests==2.32.3
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10

sqlparse==0.5.3
tzdata==2025.2
